Rewrite `curry` function generator
This generator has been updated to Swift 2.0 and also has been rewritten
to generate endless `curry` functions by concatenating the alphabet as
it goes, ie ...Y, Z, AA, AB, etc.

The biggest change is that it was rewritten to use the other curry
syntax, `(a: A)(b: B)` instead of `A -> B`. This is easier on the
compiler and we can far extend the limit. It also compiles instantly.

The limit of generic parameters was found to be 64, which I think is
interestingly a power of 2. I wonder if that number is either hard
coded in Swift or if Swift internally uses a small amount of bits for
this. ANYWAYS, I think if you need more than a 64 parameter `curry`
function then you're probably doing it wrong.