Revert to explicit closure syntax
The fancy implicit curried syntax is being removed in Swift 3.0, so we
should get ahead of the curve here and revert back to using an explicit
closure syntax.

The biggest issue with this change is that we're now limited in the
number of implementations we can provide. Specifically, a single
implementation takes a concrete amount of time to compile, but that
compile time becomes exponential as you go deeper. We waited for over a
minute to try to compile the result of `bin/generate 27` and it still
hadn't completed. 20 seems to be a decent compromise (takes about 8
seconds to compile on my machine)

This will constitute a breaking change for users, since we are removing
roughly 40 implementations of `curry`.