Remove backticks
The backticks were initially added because we were hitting `as` in our
generic parameters and needed to avoid the keyword. We have since
changed the way we're generating curry functions and can no longer
generate so many. Since we won't hit `as` we can remove the backticks.
It also seems to have improved compile duration by a few seconds
although this was based off of counting in my head.